feature/move menu class
