+++ estoquecerto/app/src/main/AndroidManifest.xml
@@ -17,7 +17,7 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
-            android:name=".DrawerBaseActivity"
+            android:name=".activities.menu.DrawerBaseActivity"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
