+++ estoquecerto/app/src/main/java/com/sollute/estoque_certo/activities/client/NewClientActivity.kt
@@ -3,7 +3,7 @@ package com.sollute.estoque_certo.activities.client
 import android.content.Intent
 import android.os.Bundle
 import android.widget.Toast
-import com.sollute.estoque_certo.DrawerBaseActivity
+import com.sollute.estoque_certo.activities.menu.DrawerBaseActivity
 import com.sollute.estoque_certo.databinding.ActivityNewClientBinding
 import com.sollute.estoque_certo.models.client.NewClient
 import com.sollute.estoque_certo.rest.Rest
