+++ estoquecerto/app/src/main/java/com/sollute/estoque_certo/activities/employee/EditEmployeeActivity.kt
@@ -4,7 +4,7 @@ import android.app.AlertDialog
 import android.content.Intent
 import android.os.Bundle
 import android.widget.Toast
-import com.sollute.estoque_certo.DrawerBaseActivity
+import com.sollute.estoque_certo.activities.menu.DrawerBaseActivity
 import com.sollute.estoque_certo.databinding.ActivityEditEmployeeBinding
 import com.sollute.estoque_certo.models.employee.EditEmployee
 import com.sollute.estoque_certo.rest.Rest
