+++ estoquecerto/app/src/main/java/com/sollute/estoque_certo/activities/employee/EmployeeActivity.kt
@@ -2,7 +2,7 @@ package com.sollute.estoque_certo.activities.employee
 
 import android.content.Intent
 import android.os.Bundle
-import com.sollute.estoque_certo.DrawerBaseActivity
+import com.sollute.estoque_certo.activities.menu.DrawerBaseActivity
 import com.sollute.estoque_certo.activities.extract.ExtractActivity
 import com.sollute.estoque_certo.activities.product.ProductActivity
 import com.sollute.estoque_certo.databinding.ActivityEmployeeBinding
