+++ estoquecerto/app/src/main/java/com/sollute/estoque_certo/activities/product/NewProductSecondActivity.kt
@@ -4,7 +4,7 @@ import android.content.Intent
 import android.os.Bundle
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
-import com.sollute.estoque_certo.DrawerBaseActivity
+import com.sollute.estoque_certo.activities.menu.DrawerBaseActivity
 import com.sollute.estoque_certo.databinding.ActivityNewProductSecondBinding
 import com.sollute.estoque_certo.models.product.NewProduct
 import com.sollute.estoque_certo.rest.Rest
