+++ estoquecerto/app/src/main/java/com/sollute/estoque_certo/activities/provider/EditProviderActivity.kt
@@ -4,7 +4,7 @@ import android.app.AlertDialog
 import android.content.Intent
 import android.os.Bundle
 import android.widget.Toast
-import com.sollute.estoque_certo.DrawerBaseActivity
+import com.sollute.estoque_certo.activities.menu.DrawerBaseActivity
 import com.sollute.estoque_certo.activities.client.ClientActivity
 import com.sollute.estoque_certo.databinding.ActivityEditProviderBinding
 import com.sollute.estoque_certo.models.provider.EditProvider
