+++ estoquecerto/app/src/main/java/com/sollute/estoque_certo/rest/Rest.kt
@@ -19,5 +19,4 @@ object Rest {
             .addConverterFactory(GsonConverterFactory.create())
             .build()
     }
-
 }
