+++ estoquecerto/app/src/main/res/layout/activity_drawer_base.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".DrawerBaseActivity"
+    tools:context=".activities.menu.DrawerBaseActivity"
     tools:openDrawer="start">
 
     <include
