+++ estoquecerto/app/src/main/res/values/strings.xml
@@ -14,6 +14,9 @@
     <string name="no">Não</string>
     <string name="terms">Eu li e concordo com os termos de uso.</string>
 
+
+
+
     <!--  Produto  -->
     <string name="product_name">Produtos</string>
     <string name="product_new">Novo Produto</string>
